Remove auth module
