+++ server/src/app.module.ts
@@ -2,8 +2,7 @@ import { Module } from "@nestjs/common";
 import { MongooseModule } from "@nestjs/mongoose";
 import { AppController } from "./app.controller";
 import { Config } from "./config";
-import { UsersModule } from "./users/users.module";
-import { AuthModule } from "./auth/auth.module";
+import { UsersModule } from "./user/users.module";
 
 @Module({
   imports: [
@@ -11,7 +10,6 @@ import { AuthModule } from "./auth/auth.module";
       `mongodb://${Config.MONGO_ROOT_USERNAME}:${Config.MONGO_ROOT_PASSWORD}@db`
     ),
     UsersModule,
-    AuthModule,
   ],
   controllers: [AppController],
 })
