--- server/src/auth/auth.module.ts
@@ -1,9 +1,0 @@
-import { forwardRef, Module } from "@nestjs/common";
-import { AuthController } from "./auth.controller";
-import { UsersModule } from "../users/users.module";
-
-@Module({
-  imports: [forwardRef(() => UsersModule)],
-  controllers: [AuthController],
-})
-export class AuthModule {}
