+++ server/src/user/users.controller.ts
@@ -10,9 +10,12 @@ import { RegisterRequestDto } from "./dto/register.dto";
 import { User } from "./schemas/user.schema";
 import { UsersService } from "./users.service";
 import { ApiResponse, ApiTags } from "@nestjs/swagger";
+import { LoginRequestDto, LoginResponseDto } from "./dto/login.dto";
+import comparePassword from "./util/comparePassword";
+import issueToken from "./util/issueToken";
 
-@ApiTags("users")
-@Controller("users")
+@ApiTags("user")
+@Controller("user")
 export class UsersController {
   constructor(private readonly usersService: UsersService) {}
 
@@ -44,6 +47,45 @@ export class UsersController {
 
     await this.usersService.create(createUserDto);
     console.log("User created");
+  }
+
+  @ApiResponse({
+    status: 200,
+    description: "Successful login",
+    type: LoginResponseDto,
+  })
+  @ApiResponse({
+    status: HttpStatus.UNAUTHORIZED,
+    description: "Invalid id or password",
+  })
+  @Post("login")
+  async login(@Body() request: LoginRequestDto) {
+    const user = await this.usersService.findOne({
+      id: request.id,
+    });
+    if (!user) {
+      throw new HttpException(
+        "Invalid id or password",
+        HttpStatus.UNAUTHORIZED
+      );
+    }
+
+    const passwordMatched = await comparePassword(
+      request.password,
+      user.hashedPassword
+    );
+    if (!passwordMatched) {
+      throw new HttpException(
+        "Invalid id or password",
+        HttpStatus.UNAUTHORIZED
+      );
+    }
+
+    const accessToken = await issueToken(user);
+
+    return {
+      accessToken,
+    } as LoginResponseDto;
   }
 
   @Get()
