+++ server/src/user/util/validateToken.ts
@@ -4,8 +4,9 @@ import { Config } from "../../config";
 export default async function validateToken(token: string) {
   const payload = jwt.verify(token, Config.JWT_SECRET, {});
   let userId = (payload as any)["userId"];
-  if (typeof userId !== "string") {
-    return { userId };
+  if (typeof userId === "string") {
+    return userId;
+  } else {
+    return null;
   }
-  throw new Error("Invalid token payload");
 }
